Move new validation into model directory
